% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{abc_control.RABC}
\alias{abc_control.RABC}
\title{abc_control.RABC}
\usage{
abc_control.RABC(n = 1000, a = 0.5, c1 = 0.01, R = 10,
  eps_final = 0, pacc_final = 0.02, d_eps_final = 0,
  num_drop = ceiling(n * a), num_keep = n - num_drop, n_param = NA,
  cov_func = ifelse(n_param == 1, stats::var, stats::cov),
  prior_eval = function(x, y) {     return(1) })
}
\arguments{
\item{n}{numeric number of proposals in output}

\item{a}{numeric fraction of proposals to drop at each iteration}

\item{c1}{controls number of MCMC steps to make at each iteration}

\item{R}{starting value of number of MCMC steps to make at each iteration}

\item{eps_final}{numeric algorithm will finish when proposals have distances below this number}

\item{pacc_final}{numeric algorithm will finish when acceptance proportion is below this number}

\item{d_eps_final}{numeric algorithm will stop when percent difference in epsilon between iterations is below this number}

\item{num_drop}{numeric number of proposals to drop at each iteration}

\item{num_keep}{numeric number of proposals to keep at each iteration}

\item{n_param}{numeric for internal use, don't change}

\item{cov_func}{function method for computing covariance}

\item{prior_eval}{evaluator of prior function}
}
\description{
Create list of control parameters for Replenishment ABC sampler
}
