% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_core.R
\name{abc_start}
\alias{abc_start}
\title{ABC algorithm}
\usage{
abc_start(prior, distance, transition = NA, method = "rejection",
  control = list(), parallel = FALSE, cl = NULL)
}
\arguments{
\item{prior}{function for prior}

\item{distance}{function for distance comp}

\item{transition}{function for transition}

\item{method}{algorithm to use}

\item{options}{list of options to use in algorithm}
}
\description{
ABC algorithm
}
\examples{
prior <- function(n){data.frame(mean = rnorm(n, 5))}
distance <- function(x){
  data <- rnorm(1000, x)
  output <- abs(mean(data) - 3)
  return(output)
}

abc_post_1 <- abc_start(prior, distance, method = "rejection", parallel = FALSE, control = list(epsilon = 0.1))

hist(abc_post_1$mean)

prior <- function(n){data.frame(mean = runif(n, 2, 4), sd = rgamma(n, 1, 1))}
distance <- function(x){
  data <- rnorm(1000, x[1], x[2])
  output <- sqrt( (mean(data) - 3)^2 + (sd(data) - 1)^2)
  return(output)
}

abc_post_2 <- abc_start(prior, distance, method = "rejection", parallel = FALSE, control = list(epsilon = 0.1))

hist(abc_post_2$mean)
hist(abc_post_2$mean)

}
