% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fun.R
\name{prior_unif}
\alias{prior_unif}
\title{Make a uniform prior}
\usage{
prior_unif(lhs_lim, rhs_lim, var_names = NULL, eval = FALSE)
}
\arguments{
\item{lhs_lim}{vector of lower bounds for uniform prior}

\item{rhs_lim}{vector of upper bounds for uniform prior}

\item{var_names}{optional character vector of parameter names}

\item{eval}{boolean if FALSE outputs a prior sampler function (similar to \code{runif}), otherwise outputs an evaluation function (similar to \code{dunif}).}
}
\description{
Make a uniform prior
}
\examples{
prior <- prior_unif(c(0, 0.5), c(1, 1), var_names = c("mu", "sd"))
prior_eval <- prior_unif(c(0, 0.5), c(1, 1), var_names = c("mu", "sd"), eval = TRUE)

sample <- rnorm(1000, mean = 0.2, sd = 0.7)
inp <- list(
    sample_mean = mean(sample),
    sample_sd = sd(sample)
)


distance <- function(theta, inp){

  sim <- rnorm(1000, theta["mu"], theta["sd"])

  output <- sqrt( (mean(sim) - inp$sample_mean)^2 + (sd(sim) - inp$sample_sd)^2)
  return(output)
}

abc_post <- abc_start(
  prior,
  distance,
  inp,
  method = "RABC",
  control = list(n = 20, prior_eval = prior_eval, pacc_final = 0.1),
  output_control = list(print_output = FALSE)
)
}
