% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fun.R
\name{prior_unif}
\alias{prior_unif}
\title{Make a uniform prior}
\usage{
prior_unif(lhs_lim, rhs_lim, var_names = NULL, eval = FALSE)
}
\arguments{
\item{lhs_lim}{vector}

\item{rhs_lim}{vector}

\item{eval}{boolean}
}
\description{
Make a uniform prior
}
